Move sql into models
